Add a note about dark mode
After the first CSS step, iOS devices in "dark mode" may not be able to see the "Email Address" and "Password" field hints until the next step when the core theme is applied. I am unsure whether this is an issue in NativeScript that needs fixing or not, but for now, I would like to add this note to the tutorial to help other potentially confused learners.