Correct deprecated use of frameModule.topmost()
In the previous step(s) this was already corrected, but this step introduces a deprecated call "frameModule.topmost()" which will result in a deprecation warning being output to the playground console. Instead, use "frameModule.Frame.topmost()" (as in prior steps) which will avoid generating this warning. (Unless you want to add this as an "extra credit" style fix for playground learners...)